Начальный коммит.

+++ Source/External/Luabind/implicit_cast.cpp
@@ -19,13 +19,10 @@
 // ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
 // OR OTHER DEALINGS IN THE SOFTWARE.
-
 #include "pch.h"
-
-#include "luabind/lua_include.hpp"
-
-#include "luabind/luabind.hpp"
-#include "luabind/detail/implicit_cast.hpp"
+#include <luabind/lua_include.hpp>
+#include <luabind/luabind.hpp>
+#include <luabind/detail/implicit_cast.hpp>
 
 namespace luabind { namespace detail {
 
@@ -46,9 +43,9 @@ namespace luabind { namespace detail {
     int implicit_cast(
         class_rep const* crep
       , LUABIND_TYPE_INFO const& type_id
-      , int& pointer_offset)
+      , ptrdiff_t& pointer_offset)
     {
-        int offset = 0;
+        ptrdiff_t offset = 0;
         if (LUABIND_TYPE_INFO_EQUAL(crep->type(), type_id)) return 0;
 
         for (vector_class<class_rep::base_info>::const_iterator i =
